pyproject.toml
[project]
name = "api-personal-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.7.0",
    "pyjwt>=2.10.1",
    "robyn>=0.64.2",
    "ruff>=0.8.3",
    "sqlalchemy-libsql>=0.1.0",
]
